requirements.txt
Django~=5.2.1
gunicorn
dotenv~=0.9.9
psycopg2-binary
djangorestframework~=3.16.0
python-dotenv~=1.1.0